Remove Sudo command
If the package is not yet installed on the system, this is not gonna work.
The solution was removing the sudo command, or adding sudo in the package installation.
Adding sudo in the package installation is useless because they need root for doing the package installation, so there is no point for sudo.

That was just a copy-paste without thinking :)